+++ src/admin_ui/src/resources.tsx
@@ -14,12 +14,14 @@ import {
 } from "react-admin";
 import {IApiInfo, IResourceInfo} from "./apiInfo";
 import {getFields, getInputs} from "./fields";
+import {makeStyles} from '@material-ui/core';
 
 
 export function getResources(apiInfo: IApiInfo) {
     return (
         apiInfo.mapResources((resourceInfo) => (
             <Resource
+                key={resourceInfo.name}
                 name={resourceInfo.name}
                 {...getResourceViews(resourceInfo)}
             />
@@ -38,18 +40,25 @@ function getResourceViews(resourceInfo: IResourceInfo) {
 }
 
 
+const useGridStyles = makeStyles({
+    headerCell: {
+        backgroundColor: '#e0e0e0',
+    },
+});
+
+
 function getListView(resourceInfo: IResourceInfo) {
     const {list, show, edit} = resourceInfo.views;
     if (!list) {
         return null;
     }
-
     return (props: any) => {
         if (!resourceInfo.deletable)
             props['bulkActionButtons'] = false;
+        const classes = useGridStyles();
         return (
             <List {...props}>
-                <Datagrid rowClick={show ? 'show' : ''}>
+                <Datagrid rowClick={show ? 'show' : ''} classes={classes}>
                     {getFields(list.fields)}
                     {edit ? <EditButton/> : null}
                 </Datagrid>
@@ -96,7 +105,7 @@ function getCreateView(resourceInfo: IResourceInfo) {
 
 const EditWithoutDeleteToolbar = props => (
     <Toolbar {...props} >
-        <SaveButton />
+        <SaveButton/>
     </Toolbar>
 );
 
@@ -108,7 +117,7 @@ function getEditView(resourceInfo: IResourceInfo) {
     }
     let form_props = {};
     if (!resourceInfo.deletable) {
-        form_props['toolbar'] = <EditWithoutDeleteToolbar />;
+        form_props['toolbar'] = <EditWithoutDeleteToolbar/>;
     }
 
     return (props: any) => {