+++ src/example_app/example_app/users/resources.py
@@ -4,10 +4,10 @@
 :Date: 05.02.2020
 """
 from dataclasses import asdict, dataclass
-from typing import Dict, List, Literal, Optional
+from typing import Dict, List, Literal, Optional, Union
 
 import pendulum
-from pyramid.security import Allow, Authenticated, DENY_ALL, Everyone
+from pyramid.security import Allow, Everyone
 from restfw.hal import HalResource, HalResourceWithEmbedded, list_to_embedded_resources
 from restfw.interfaces import MethodOptions
 from restfw.schemas import GetEmbeddedSchema
@@ -23,6 +23,12 @@ class ChildModel:
 
 
 @dataclass()
+class WorkModel:
+    title: str
+    address: str
+
+
+@dataclass()
 class UserModel:
     id: int
     created: pendulum.DateTime
@@ -31,6 +37,7 @@ class UserModel:
     age: int
     sex: Optional[Literal['m', 'f']]
     children: List[ChildModel]
+    current_work: WorkModel
 
 
 class User(HalResource):
@@ -105,8 +112,13 @@ class Users(HalResourceWithEmbedded):
     def get_user_by_model(self, model: UserModel):
         return User(model, parent=self)
 
-    def create_user(self, first_name, last_name, age: Optional[int] = None, sex: Optional[Literal['m', 'f']] = 'm'):
+    def create_user(self, first_name, last_name, age: Optional[int] = None, sex: Optional[Literal['m', 'f']] = 'm',
+                    current_work: Union[None, dict, WorkModel] = None):
         user_id = self._next_id
+        if not current_work:
+            current_work = WorkModel(title='', address='')
+        if isinstance(current_work, dict):
+            current_work = WorkModel(**current_work)
         model = UserModel(
             id=user_id,
             created=pendulum.now(pendulum.UTC),
@@ -115,6 +127,7 @@ class Users(HalResourceWithEmbedded):
             age=age,
             sex=sex,
             children=[],
+            current_work=current_work,
         )
         self._next_id += 1
         self._models[str(user_id)] = model